Added the main installation page.
Updated to .Net Core 2.1, packages, and views.

+++ showcase/Services/EmailSender.cs
@@ -25,7 +25,7 @@ namespace showcase.Services
                 UseDefaultCredentials = false,
                 DeliveryMethod = SmtpDeliveryMethod.Network,
                 Host = smtpServer.Host,
-                Port = smtpServer.Port,
+                Port = (int)smtpServer.Port,
                 EnableSsl = true,
                 Credentials = new NetworkCredential(smtpServer.Username, smtpServer.Password)
             };
